build: split cog services on spark tests
