+++ pipeline.yaml
@@ -266,8 +266,11 @@ jobs:
         PACKAGE: "automl"
       cntk:
         PACKAGE: "cntk"
-      cognitive:
-        PACKAGE: "cognitive"
+      cognitive1:
+        PACKAGE: "cognitive.split1"
+        FLAKY: "true"
+      cognitive2:
+        PACKAGE: "cognitive.split2"
         FFMPEG: "true"
         FLAKY: "true"
       core:
@@ -326,7 +329,7 @@ jobs:
         azureSubscription: 'MMLSpark Connection'
         scriptLocation: inlineScript
         inlineScript: |
-          (${FFMPEG:-false} && sudo apt-get install ffmpeg libgstreamer1.0-0 \
+          (${FFMPEG:-false} && sudo apt-get update && sudo apt-get install ffmpeg libgstreamer1.0-0 \
           gstreamer1.0-plugins-base gstreamer1.0-plugins-good gstreamer1.0-plugins-bad gstreamer1.0-plugins-ugly -y)
           export SBT_OPTS="-Xmx2G -XX:+UseConcMarkSweepGC -XX:+CMSClassUnloadingEnabled -XX:MaxPermSize=2G -Xss2M  -Duser.timezone=GMT"
           (timeout 20m sbt coverage "testOnly com.microsoft.ml.spark.$(PACKAGE).**") ||
