+++ src/test/scala/com/microsoft/ml/spark/cognitive/split1/FaceSuite.scala
@@ -1,12 +1,13 @@
 // Copyright (C) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in project root for information.
 
-package com.microsoft.ml.spark.cognitive
+package com.microsoft.ml.spark.cognitive.split1
 
 import java.util.UUID
 
 import com.microsoft.ml.spark.Secrets
 import com.microsoft.ml.spark.cognitive.cognitive.Face
+import com.microsoft.ml.spark.cognitive._
 import com.microsoft.ml.spark.core.test.fuzzing.{TestObject, TransformerFuzzing}
 import org.apache.spark.ml.util.MLReadable
 import org.apache.spark.sql.functions.{col, lit}
